+++ test/src/test/java/org/metavm/racecondition/RaceConditionTest.java
@@ -36,7 +36,7 @@ public class RaceConditionTest extends TestCase {
     public void test() throws InterruptedException {
         MockUtils.assemble("/Users/leen/workspace/object/test/src/test/resources/asm/race_condition.masm",
                 typeManager, schedulerAndWorker);
-        long size = (long) TestUtils.doInTransaction(() -> apiClient.callMethod("Utils", "size", List.of(
+        long size = (int) TestUtils.doInTransaction(() -> apiClient.callMethod("Utils", "size", List.of(
                 List.of()
         )));
         Assert.assertEquals(0, size);
